Publish binaries with node pre-gyp
This uses the binaries built within the docker image build on the
testing step to package with the release for use with node pre-gyp.

Node pre-gyp will look for published artifacts for the given
architecture before trying to build the module locally. This should
allow projects relying on this dependency to build much faster,
particularly on constrained environments.

Change-type: minor